Updated docmentation and began moving type definitions to components. Also deleted unneeded files
