+++ src/_assets/icons.ts
@@ -1,15 +1,23 @@
+/**
+ * Collection of icons used throughout the application
+ * @module icons
+ */
 import { ArrowPathIcon } from "@heroicons/react/16/solid";
+/** Icon for reset functionality */
 export const ResetIcon = ArrowPathIcon;
 
 import { BsKeyboard } from "react-icons/bs";
+/** Icon representing QWERTY keyboard input */
 export const QwertyIcon = BsKeyboard;
 
 import { TbUpload } from "react-icons/tb";
+/** Icon for file upload functionality */
 export const UploadFileIcon = TbUpload;
 
-import { HiPlayCircle } from "react-icons/hi2";
-import { HiPauseCircle } from "react-icons/hi2";
-import { HiStopCircle } from "react-icons/hi2";
+import { HiPlayCircle, HiPauseCircle, HiStopCircle } from "react-icons/hi2";
+/** Icon for play functionality */
 export const PlayIcon = HiPlayCircle;
+/** Icon for pause functionality */
 export const PauseIcon = HiPauseCircle;
+/** Icon for stop functionality */
 export const StopIcon = HiStopCircle;
