+++ src/_components/optionsPanel/midi/MidiPlayback.tsx
@@ -1,12 +1,32 @@
-import { MidiPlaybackProps } from "@/_lib/_types/types";
 import { UploadFileIcon } from "@/_assets/icons";
-import { PlayIcon, PauseIcon, StopIcon } from "@/_assets/icons";
-import MidiPlaybackButton from "./MidiPlaybackButton"
+import MidiPlaybackButton from "./MidiPlaybackButton";
+
+export interface MidiPlaybackProps {
+    midiFileText: string | null;
+    handleMidiUpload: (e: React.ChangeEvent<HTMLInputElement>) => void;
+    play: () => void; 
+    pause: () => void;
+    stop: () => void;
+    playbackState:  'stopped' | 'playing' | 'paused';
+};
 
 /**
- * MidiPlayback component provides controls for MIDI file playback and upload functionality
- * @component
-*/
+ * Props for MIDI playback controls
+ * @interface MidiPlaybackProps
+ * @property {string | null} midiFileText - Display text for the currently loaded MIDI file
+ * @property {(e: React.ChangeEvent<HTMLInputElement>) => void} handleMidiUpload - Handler for MIDI file upload events
+ * @property {() => void} play - Function to start MIDI playback
+ * @property {() => void} pause - Function to pause MIDI playback
+ * @property {() => void} stop - Function to stop MIDI playback
+ * @property {'stopped' | 'playing' | 'paused'} playbackState - Current state of MIDI playback
+ */
+
+/**
+ * Component providing MIDI file playback controls and file upload functionality
+ * Displays the current MIDI file name and playback state
+ * @param {MidiPlaybackProps} props - The component props
+ * @returns {JSX.Element} Rendered MIDI playback controls
+ */
 const MidiPlayback: React.FC<MidiPlaybackProps> = ({
     midiFileText, 
     handleMidiUpload,
@@ -14,7 +34,7 @@ const MidiPlayback: React.FC<MidiPlaybackProps> = ({
     pause, 
     stop, 
     playbackState
-  }) => {
+  }: MidiPlaybackProps): JSX.Element => {
 
     return (
         <div className="gap-2 p-2 bg-cyan-100 rounded-lg m-2 grid grid-flow-row">
