+++ src/_hooks/useQwertyInput.ts
@@ -1,6 +1,15 @@
 import { useCallback, useEffect, useRef, useState } from "react";
-import { QwertyInputProps } from "@/_lib/_types/types";
+import { KeyHandlers } from "@/_lib/_types/types";
 import { qwertyMap } from "@/_lib/_data/qwertyMap";
+
+export interface QwertyInputProps {
+    octaveRange: {
+        octaveMin: number;
+        currentOctave: number;
+        octaveMax: number; 
+    };
+    keyHandlers: KeyHandlers;
+};
 
 interface QwertyInputReturn {
     isQwertyEnabled: boolean;
@@ -8,16 +17,22 @@ interface QwertyInputReturn {
 }
 
 /**
- * Custom hook that enables QWERTY keyboard input for musical applications.
- * Maps keyboard keys to musical notes and handles octave switching.
+ * Custom hook for handling QWERTY keyboard input in a musical application
+ * Provides keyboard-to-note mapping and octave management functionality
  * 
- * @param {QwertyInputProps} - The hook's configuration object
- * @param {OctaveRange} QwertyInputProps.octaveRange - Object containing octave range constraints
- * @param {KeyHandlers} QwertyInputProps.keyHandlers - Object containing key event callbacks
+ * @param {Object} props - Hook configuration object
+ * @param {Object} props.octaveRange - Configuration for octave boundaries and current position
+ * @param {number} props.octaveRange.octaveMin - Lowest allowable octave number
+ * @param {number} props.octaveRange.currentOctave - Currently selected octave
+ * @param {number} props.octaveRange.octaveMax - Highest allowable octave number
+ * @param {Object} props.keyHandlers - Event handlers for key interactions
+ * @param {(noteName: string) => void} props.keyHandlers.onKeyDown - Callback triggered when a key is pressed
+ * @param {(noteName: string) => void} props.keyHandlers.onKeyUp - Callback triggered when a key is released
  * 
- * @returns {QwertyInputProps} An object containing:
- *   - isQwertyEnabled: boolean indicating if QWERTY input is active
- *   - toggleIsQwertyEnabled: function to toggle QWERTY input on/off
+ * @returns {Object} Hook control interface
+ * @returns {boolean} returns.isQwertyEnabled - Current state of QWERTY input
+ * @returns {() => void} returns.toggleIsQwertyEnabled - Function to toggle QWERTY input on/off
+ * 
  * 
  * @example
  * const { isQwertyEnabled, toggleIsQwertyEnabled } = useQwertyInput({
