+++ src/_lib/_data/qwertyMap.ts
@@ -1,5 +1,33 @@
-import { QwertyMap } from "../_types/types";
+/**
+ * Mapping interface for QWERTY keyboard keys to musical notes
+ * Associates keyboard characters with their corresponding musical pitch and octave
+ * 
+ * @interface QwertyMap
+ * @property {BaseKeyName} [key: string] - Maps each keyboard character to its musical properties
+ */
+interface QwertyMap {
+    [key: string]: BaseKeyName;
+};
 
+/**
+ * Properties defining a musical note's basic characteristics
+ * 
+ * @interface BaseKeyName
+ * @property {string} pitch - The musical pitch (e.g., 'C', 'C#', 'D', etc.)
+ * @property {number} baseOctave - The octave offset (0 for lower octave, 1 for higher octave)
+ */
+interface BaseKeyName {
+    pitch: string;
+    baseOctave: number;
+};
+
+/**
+ * Standard QWERTY keyboard mapping for a two-octave piano layout
+ * Maps keyboard keys to corresponding musical notes in a piano-like arrangement
+ * 
+ * @constant
+ * @type {QwertyMap}
+ */
 export const qwertyMap: QwertyMap = {
 	'a' : {
 		pitch: 'C',
