+++ src/_lib/_types/types.ts
@@ -1,46 +1,3 @@
-// Options Panel Types
-export interface OptionsPanelProps {
-    globalProps: GlobalProps;
-    inputProps: InputProps;
-    modeProps: ModeProps;
-    outputProps: OutputProps;
-};
-
-export interface GlobalProps {
-
-};
-
-export interface InputProps {
-    isQwertyEnabled: boolean;
-    toggleIsQwertyEnabled: () => void;
-    midiPlaybackProps: MidiPlaybackProps;
-};
-
-export interface MidiPlaybackProps {
-    midiFileText: string | null;
-    handleMidiUpload: (e: React.ChangeEvent<HTMLInputElement>) => void;
-    play: () => void; 
-    pause: () => void;
-    stop: () => void;
-    playbackState:  'stopped' | 'playing' | 'paused';
-};
-
-export interface ModeProps {
-    getModeRef: () => Mode;
-    setModeRef: (newMode: Mode) => void;
-    updateModifier: (newValue: number, index: number) => void;
-    modes: Mode[];
-    onReset: () => void;
-};
-
-export interface OutputProps {
-    polySynthInterface: EffectInterface;
-    effectsInterfaces: {
-        reverbInterface: EffectInterface;
-        vibratoInterface: EffectInterface;
-    };
-};
-
 // Mode Types
 export interface Mode {
     id: string;
@@ -59,20 +16,6 @@ export interface ModeModifiers {
     step: number;
 }
 
-//Keyboard Types
-export interface KeyboardProps {
-    keys: Key[],
-    keyEmitter: ToneType.Emitter | null,
-    keyHandlers: KeyHandlers;
-}
-
-export interface Key {
-    midiNumber: number;
-    name: string;
-    pitch: string;
-    octave: number;
-};
-
 export interface KeyboardRange {
     startPitch: string; 
     startOctave: number; 
@@ -81,11 +24,11 @@ export interface KeyboardRange {
 };
 
 // Hertz State Types
-export interface ModeSelect {
-    mode: string;
-    hertzModifiers: HertzModifiers;
-    hertzTable: HertzTable;
-};
+// export interface ModeSelect {
+//     mode: string;
+//     hertzModifiers: HertzModifiers;
+//     hertzTable: HertzTable;
+// };
 
 export interface HertzModifiers {
     lastKey: string;
@@ -96,6 +39,13 @@ export interface HertzTable {
     [key: string]: number;
 };
 
+export interface Key {
+    midiNumber: number;
+    name: string;
+    pitch: string;
+    octave: number;
+};
+
 // Key Handler Types
 export interface KeyHandlers {
     onKeyDown: (keyName: string) => void;
@@ -103,37 +53,22 @@ export interface KeyHandlers {
 }
 
 // User Input Types
-export interface QwertyInputProps {
-    octaveRange: OctaveRange;
-    keyHandlers: KeyHandlers;
-};
+// export interface QwertyInputProps {
+//     octaveRange: OctaveRange;
+//     keyHandlers: KeyHandlers;
+// };
 
-interface OctaveRange {
-    octaveMin: number;
-    currentOctave: number;
-    octaveMax: number; 
-};
-
-export interface QwertyMap {
-    [key: string]: BaseKeyName;
-};
-
-export interface BaseKeyName {
-    pitch: string;
-    baseOctave: number;
-};
+// interface OctaveRange {
+//     octaveMin: number;
+//     currentOctave: number;
+//     octaveMax: number; 
+// };
 
 export type MidiPlaybackState = 'stopped' | 'playing' | 'paused';
 
 // Audio Output Types
 import * as ToneType from "tone";
 export { ToneType };
-
-
-export interface Note {
-    keyName: string;
-    hertz: number;
-};
 
 export interface HertzPlayback {
     playHertz: (keyName: string, hertz:number) => void;
